add new folders for different components and change app.tsx
