+++ src/components/landing/about.tsx
