+++ src/pages/AboutPage.tsx
@@ -1,8 +1,8 @@
 import Navbar from "../components/navbar";
-import HeroSectionAbout from "../components/HeroSectionAbout";
+import HeroSectionAbout from "../components/about/HeroSectionAbout";
 import Footer from "../components/footer";
-import AboutUs from "../components/AboutUs";
-import ContactUs from "../components/ContactUs";
+import AboutUs from "../components/about/AboutUs";
+import ContactUs from "../components/about/ContactUs";
 export default function AboutContact() {
   return (
     <>
