+++ src/pages/landing.tsx
@@ -1,9 +1,9 @@
 import Navbar from "../components/navbar";
-import HeroSection from "../components/heroSection";
+import HeroSection from "../components/landing/heroSection";
 import Footer from "../components/footer";
-import Services from "../components/featureSection";
-import TutorialSection from "../components/tutorial";
-import About from "../components/about";
+import Services from "../components/landing/featureSection";
+import TutorialSection from "../components/landing/tutorial";
+import About from "../components/landing/about";
 
 export default function Landing() {
   return (
